Fix: Convert file references to project
found another common problem. Lots of our references that are file based should actually be project references. This item still needs unit tests. I started them, realized that I didn't really know what the shape of a field was, ran it in the debugger to see, and ended up finishing it in one pass.