Fix for reference version
The include statement captures the version number of references. This fix checks the file at the hint path (if it can) and then updates the version number of the reference to match. It's mostly pointless, but in our case we have a 0.0.0.0 file floating around on the build server, when 1.0.0.0 is current, and lots of teams referencing 0.0.0.0.